Refactor + Added bootstrap dependency
