+++ src/Field.js
@@ -1,0 +1,8 @@
+const Field = ({name, children}) =>
+  <form>
+    <input
+      id={name}
+      type="text"
+      placeholder = {children}
+      />
+  </form>
