+++ src/Table.js
@@ -1,0 +1,20 @@
+const Table = ({ list, pattern, onDismiss}) =>
+
+  <div className="table">
+      { list.filter(isSearched(pattern)).map(item =>
+        <div key={item.objectID} className="table-row">
+          <span style={{ width: '30%' }}>{item.genericName}</span>
+          <span style={{ width: '30%' }}>{item.brandName}</span>
+          <span style={{ width: '10%' }}>{item.dosage}</span>
+          
+          <span style={{ width: '10%' }}>
+          <Button 
+            onClick={() => onDismiss(item.objectID)}
+            className="button-inline"
+          >
+            Dismiss
+          </Button>
+          </span>
+        </div>
+      )}
+    </div>
